8.2.11. создает детальную карточку предложения, делает переходы с нее с главной страницы, со страницы оффера и со страницы избранного. Рефакторит код.
